Split out the actual FFI from raw_ffi
This commit splits up x11rb::xcb_ffi:raw_ffi so that raw_ffi now
contains the common struct definitions and a new ffi sub-module contains
the actual code for calling into libxcb. This is mostly the
implementation of the dl-libxcb feature.
